requirements.txt
ipykernel
python-dotenv
owlready2
EMMOntoPy
pydotplus
ipycytoscape
networkx
graphviz
openai